refactor(web): split types/index.ts into domain modules behind a barrel
Move the 216 exported types out of the 2578-line web/src/types/index.ts
into 17 domain modules (app, auth, automation, backups, common, crossseed,
dashboard, dir-scan, external-programs, indexers, instances, logs,
notifications, orphan-scan, rss, torrent-creation, torrents) and turn
index.ts into a thin re-export barrel. Definitions are moved verbatim — no
behavior or type changes. The barrel also folds in the pre-existing arr.ts,
so its types are now importable from "@/types" as well.

Every symbol previously exported from "@/types" remains exported, verified
by diffing the export set against develop (216 preserved, 0 missing, the
only additions being the 12 arr types). tsc, lint, vitest, and build stay
green. Existing import sites are untouched — the barrel keeps them working.

Part of #1930 (1930-b).
