+++ web/src/types/common.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2025-2026, s0up and the autobrr contributors.
+ * SPDX-License-Identifier: GPL-2.0-or-later
+ */
+
+// Generic warning response for operations that succeed with caveats
+export interface WarningResponse {
+  warning?: string
+}
